fix(text): fix stories for text component
