+++ core/components/atoms/text/__stories__/Appearance.story.jsx
@@ -1,18 +1,18 @@
 import * as React from 'react';
-import Text from '../../index';
+import { Text } from '@/index';
 
 // CSF format story
 export const appearance = () => {
   const appearances = ['default', 'white', 'destructive', 'subtle', 'disabled', 'success', 'link'];
   return (
-    <div className="d-flex">
+    <div className="d-flex w-75 justify-content-between">
       {appearances.map((appear, ind) => {
-        const backgroundClass = appear === 'white' ? 'bg-dark mr-6' : 'bg-transparent mr-6';
+        const backgroundClass = appear === 'white' ? 'bg-dark mr-6' : '';
 
         return (
-          <div key={ind} className={backgroundClass}>
-            <Text appearance={appear}>{appear.charAt(0).toUpperCase() + appear.slice(1)}</Text>
-          </div>
+          <Text key={ind} className={backgroundClass} appearance={appear}>
+            {appear.charAt(0).toUpperCase() + appear.slice(1)}
+          </Text>
         );
       })}
     </div>
@@ -20,7 +20,7 @@ export const appearance = () => {
 };
 
 export default {
-  title: 'Components/Text/Variants/Appearance',
+  title: 'Components/Text/Appearance',
   component: Text,
   parameters: {
     docs: {
