+++ core/components/atoms/text/__stories__/index.story.jsx
@@ -3,22 +3,7 @@ import { Text } from '@/index';
 
 // CSF format story
 export const all = () => {
-  const weight = 'medium';
-  const size = 'regular';
-  const small = false;
-  const appearance = 'white';
-  const options = {
-    size,
-    weight,
-    appearance,
-    small,
-  };
-
-  return (
-    <div className={appearance === 'white' ? 'bg-dark' : 'bg-transparent'}>
-      <Text {...options}>Text component have different varients, look for options in knobs tab.</Text>
-    </div>
-  );
+  return <Text>Text component have different variants, look for options in knobs tab.</Text>;
 };
 
 export default {
